Rename one file and rmv most code from md files for maintainability :fire:
